refactor: move create alert button in Bidding closed comp
